Migrations/Database created

+++ Domain/Entities/Concrete/Vacancy.cs
@@ -20,14 +20,14 @@ public class Vacancy : BaseEntity
 
 
 
-    public int CategoryId { get; set; }
+    public long CategoryId { get; set; }
 
     [ForeignKey(nameof(CategoryId))]
     public Category Category { get; set; }
 
 
 
-    public int CompanyId { get; set; }
+    public long CompanyId { get; set; }
 
     [ForeignKey(nameof(CompanyId))]
     public Company SupplierCompany { get; set; }
